+++ spec/unit/active_attr/attributes_spec.rb
@@ -4,12 +4,14 @@ require "active_attr/attributes"
 module ActiveAttr
   describe Attributes do
     subject { model_class.new }
+    let(:last_name) { "Poweski" }
 
     let :model_class do
       Class.new do
         include InitializationVerifier
         include Attributes
-        attribute :name
+        attribute :first_name
+        attribute :last_name
         attribute :amount
 
         def self.name
@@ -24,9 +26,17 @@ module ActiveAttr
           super
         end
 
-        def initialize(name=nil)
+        def last_name=(value)
+          super(value.to_s.upcase)
+        end
+
+        def last_name
+          super || "Poweski"
+        end
+
+        def initialize(first_name=nil)
           super
-          write_attribute(:name, name)
+          write_attribute(:first_name, first_name)
         end
       end
     end
@@ -43,7 +53,7 @@ module ActiveAttr
 
     describe ".attribute" do
       it "creates an attribute with no options" do
-        model_class.attributes.should include(AttributeDefinition.new(:name))
+        model_class.attributes.should include(AttributeDefinition.new(:first_name))
       end
 
       it "returns the attribute definition" do
@@ -51,8 +61,8 @@ module ActiveAttr
       end
 
       it "defines an attribute reader that calls #attribute" do
-        subject.should_receive(:attribute).with("name")
-        subject.name
+        subject.should_receive(:attribute).with("first_name")
+        subject.first_name
       end
 
       it "defines an attribute reader that can be called via super" do
@@ -61,8 +71,8 @@ module ActiveAttr
       end
 
       it "defines an attribute writer that calls #attribute=" do
-        subject.should_receive(:attribute=).with("name", "Ben")
-        subject.name = "Ben"
+        subject.should_receive(:attribute=).with("first_name", "Ben")
+        subject.first_name = "Ben"
       end
 
       it "defines an attribute writer that can be called via super" do
@@ -93,7 +103,7 @@ module ActiveAttr
       end
 
       it "renders the attribute names in alphabetical order" do
-        model_class.inspect.should match "(amount, name)"
+        model_class.inspect.should match "(amount, first_name, last_name)"
       end
 
       it "doesn't format the inspection string for attributes if the model does not have any" do
@@ -109,7 +119,7 @@ module ActiveAttr
       end
 
       it "returns false when compared to another type" do
-        should_not == Struct.new(:attributes).new("name" => "Ben")
+        should_not == Struct.new(:attributes).new("first_name" => "Ben")
       end
     end
 
@@ -122,32 +132,23 @@ module ActiveAttr
 
       context "when an attribute is defined" do
         it "returns the key value pairs" do
-          subject.name = "Ben"
-          subject.attributes.should include("name" => "Ben")
+          subject.first_name = "Ben"
+          subject.attributes.should include("first_name" => "Ben")
         end
 
         it "returns a new Hash " do
-          subject.attributes.merge!("name" => "Bob")
-          subject.attributes.should_not include("name" => "Bob")
+          subject.attributes.merge!("first_name" => "Bob")
+          subject.attributes.should_not include("first_name" => "Bob")
         end
 
         it "returns all attributes" do
-          subject.attributes.keys.should =~ %w(amount name)
+          subject.attributes.keys.should =~ %w(amount first_name last_name)
         end
       end
 
       context "when a getter is overridden" do
-        before do
-          subject.extend Module.new {
-            def name
-              "Benjamin"
-            end
-          }
-        end
-
         it "uses the overridden implementation" do
-          subject.name = "Ben"
-          subject.attributes.should include("name" => "Benjamin")
+          subject.attributes.should include("last_name" => last_name)
         end
       end
     end
@@ -159,27 +160,19 @@ module ActiveAttr
     end
 
     describe "#inspect" do
-      before { subject.name = "Ben" }
+      before { subject.first_name = "Ben" }
 
       it "includes the class name and all attribute values in alphabetical order by attribute name" do
-        subject.inspect.should == %q{#<Foo amount: nil, name: "Ben">}
+        subject.inspect.should == %{#<Foo amount: nil, first_name: "Ben", last_name: "#{last_name}">}
       end
 
       it "doesn't format the inspection string for attributes if the model does not have any" do
-        attributeless.new.inspect.should == %q{#<Foo>}
+        attributeless.new.inspect.should == %{#<Foo>}
       end
 
       context "when a getter is overridden" do
-        before do
-          subject.extend Module.new {
-            def name
-              "Benjamin"
-            end
-          }
-        end
-
         it "uses the overridden implementation" do
-          subject.inspect.should == %q{#<Foo amount: nil, name: "Benjamin">}
+          subject.inspect.should include %{last_name: "#{last_name}"}
         end
       end
     end
@@ -188,22 +181,34 @@ module ActiveAttr
       describe "##{method}" do
         context "when an attribute is not set" do
           it "returns nil" do
-            subject.send(method, :name).should == nil
+            subject.send(method, :first_name).should be_nil
           end
         end
 
         context "when an attribute is set" do
-          let(:name) { "Bob" }
+          let(:first_name) { "Bob" }
 
-          before { subject.write_attribute(:name, name) }
+          before { subject.write_attribute(:first_name, first_name) }
 
           it "returns the attribute using a Symbol" do
-            subject.send(method, :name).should == name
+            subject.send(method, :first_name).should == first_name
           end
 
           it "returns the attribute using a String" do
-            subject.send(method, 'name').should == name
+            subject.send(method, 'first_name').should == first_name
           end
+        end
+
+        context "when the getter is overridden" do
+          it "uses the overridden implementation" do
+            subject.send(method, :last_name).should == last_name
+          end
+        end
+
+        it "raises when getting an undefined attribute" do
+          expect do
+            subject.send(method, :initials)
+          end.to raise_error UnknownAttributeError, "unknown attribute: initials"
         end
       end
     end
@@ -211,19 +216,34 @@ module ActiveAttr
     [:[]=, :write_attribute].each do |method|
       describe "##{method}" do
         it "raises ArgumentError with one argument" do
-          expect { subject.send(method, :name) }.to raise_error(ArgumentError)
+          expect { subject.send(method, :first_name) }.to raise_error(ArgumentError)
         end
 
         it "raises ArgumentError with no arguments" do
           expect { subject.send(method) }.to raise_error(ArgumentError)
         end
 
-        it "assigns sets an attribute using a Symbol and value" do
-          expect { subject.send(method, :name, "Ben") }.to change { subject.attributes["name"] }.from(nil).to("Ben")
+        it "sets an attribute using a Symbol and value" do
+          expect { subject.send(method, :first_name, "Ben") }.to change { subject.attributes["first_name"] }.from(nil).to("Ben")
         end
 
-        it "assigns sets an attribute using a String and value" do
-          expect { subject.send(method, 'name', "Ben") }.to change { subject.attributes["name"] }.from(nil).to("Ben")
+        it "sets an attribute using a String and value" do
+          expect { subject.send(method, 'first_name', "Ben") }.to change { subject.attributes["first_name"] }.from(nil).to("Ben")
+        end
+
+        it "is able to set an attribute to nil" do
+          subject.first_name = "Ben"
+          expect { subject.send(method, :first_name, nil) }.to change { subject.attributes["first_name"] }.from("Ben").to(nil)
+        end
+
+        it "uses the overridden implementation when the setter is overridden" do
+          subject.send(method, :last_name, "poweski").should == "POWESKI"
+        end
+
+        it "raises when setting an undefined attribute" do
+          expect do
+            subject.send(method, :initials, "BP")
+          end.to raise_error UnknownAttributeError, "unknown attribute: initials"
         end
       end
     end